requirements_heroku.txt
django-heroku-memcacheify==0.8
pylibmc==1.2.3